move method to utils
